Loading from asset catalog also for macOS; Refactoring
