+++ lottie-swift/src/Public/Animation/AnimationPublic.swift
@@ -80,17 +80,53 @@ public extension Animation {
       return nil
     }
   }
-  
+    
+  /**
+      Loads an animation model from the asset catalog by its name. Returns `nil` if an animation is not found.
+      - Parameter name: The name of the json file in the asset catalog. EG "StarAnimation"
+      - Parameter bundle: The bundle in which the animation is located. Defaults to `Bundle.main`
+      - Parameter animationCache: A cache for holding loaded animations. Optional.
+      - Returns: Deserialized `Animation`. Optional.
+  */
+  static func asset(_ name: String,
+                      bundle: Bundle = Bundle.main,
+                      animationCache: AnimationCacheProvider? = nil) -> Animation? {
+    /// Create a cache key for the animation.
+    let cacheKey = bundle.bundlePath + "/" + name
+
+    /// Check cache for animation
+    if let animationCache = animationCache,
+        let animation = animationCache.animation(forKey: cacheKey) {
+        /// If found, return the animation.
+        return animation
+    }
+
+    /// Load jsonData from Asset
+    guard let json = Data.jsonData(from: name, in: bundle) else {
+        return nil
+    }
+
+    do {
+      /// Decode animation.
+      let animation = try JSONDecoder().decode(Animation.self, from: json)
+      animationCache?.setAnimation(animation, forKey: cacheKey)
+      return animation
+    } catch {
+      /// Decoding error.
+      return nil
+    }
+  }
+
   /// A closure for an Animation download. The closure is passed `nil` if there was an error.
   typealias DownloadClosure = (Animation?) -> Void
-  
+
   /**
    Loads a Lottie animation asynchronously from the URL.
-   
+
    - Parameter url: The url to load the animation from.
    - Parameter closure: A closure to be called when the animation has loaded.
    - Parameter animationCache: A cache for holding loaded animations.
-   
+
    */
   static func loadedFrom(url: URL,
                                 closure: @escaping Animation.DownloadClosure,
